refactor(client.go): move Connection interface and gobConnection struct to messages/connections.go
